+++ src/@crmkit/router/RouteNames.js
@@ -1,0 +1,5 @@
+import names from "@/router/names.gen";
+
+const RouteNames = Object.freeze(names);
+
+export default RouteNames;