+++ src/@crmkit/router/router.js
@@ -1,10 +1,9 @@
 import { createRouter as _createRouter, createWebHistory } from "vue-router";
-import routes from "./routes";
+import routes from "@/router/routes";
 
 export function createRouter() {
   return _createRouter({
     history: createWebHistory(import.meta.env.BASE_URL),
-    // base: import.meta.env.BASE_URL,
     routes: routes,
   });
 }