+++ src/views/GoodsDetailsView.vue
@@ -14,6 +14,8 @@
       />
       <UISpacer />
 
+      <!-- Clone -->
+      <UIButton type="basic labeled" text="Клонировать" icon="clone" @click="cloneGood" />
       <!-- Print -->
       <UIButton type="basic labeled" text="На сборку" icon="print" @click="print" />
       <!-- Edit -->
@@ -46,16 +48,16 @@ import apiService from "@/services/api.service.js";
 import Alerts from "@/utils/alerts";
 
 import AccessRights from "@/enums/AccessRights";
+import RouteNames from "%/router/routeNames";
 
-// import RouteNames from "@/router/routeNames";
-
-import UIOptButtons from "@/components/UIOptButtons.vue";
+import UIOptButtons from "#/UIOptButtons.vue";
 
 const kTABS = [
-  { name: "ОСНОВНОЕ", id: "general" },
+  { name: "ОСНОВНОЕ", id: RouteNames.Goods.DetailsTabs.GeneralPath },
   { name: "ФОТО", id: "photos" },
   { name: "ВИДЕО", id: "videos" },
   { name: "ЗАКАЗЫ", id: "orders", access: AccessRights.OrdersView },
+  { name: "ИСТОРИЯ", id: "history" },
 ];
 
 /// SETUP
@@ -118,6 +120,14 @@ function setTitle() {
 // ---
 function edit() {
   navigateTo.Goods.Edit({ goodId: props.goodId });
+}
+
+function cloneGood() {
+  const text = "Будет создана копия этого товара. Фото и видео альбомы у этих товаров будут общие.";
+  var confirmed = confirm(text);
+  if (confirmed) {
+    postCloneGood(props.goodId);
+  }
 }
 
 async function print() {
@@ -230,5 +240,16 @@ async function postSetStatus(good_id, status) {
   }
 
   fetchGood(props.goodId);
+}
+
+async function postCloneGood(good_id, status) {
+  try {
+    let result = await apiService.cloneGood(good_id);
+
+    navigateTo.Goods.Edit({ goodId: result.id });
+    Alerts.showSuccess(`Товар клонирован!`);
+  } catch (error) {
+    Alerts.showNetworkError(error);
+  }
 }
 </script>