+++ src/views/GoodsEditView.vue
@@ -88,8 +88,6 @@ import apiService from "@/services/api.service.js";
 import { useView } from "@/composables/view";
 import { useNavigation } from "@/composables/navigation";
 
-import RouteNames from "@/router/routeNames";
-
 import GoodObject from "@/objects/Good";
 import Alerts from "@/utils/alerts";
 
@@ -132,7 +130,6 @@ export default {
       //   subTitle: "Редактирование товара",
       // },
       isLoading: false,
-      RouteNames,
     };
   },
 
@@ -181,7 +178,9 @@ export default {
           !this.good.price ||
           !this.good.quantity ||
           !this.good.brand ||
-          !this.good.description,
+          !this.good.description ||
+          this.isLoading,
+        loading: this.isLoading,
       };
     },
   },